+++ tests/src/Fixtures/Basic/PersonWithDog.php
@@ -1,0 +1,45 @@
+<?php
+
+declare(strict_types=1);
+
+/*
+ * This file is part of rekalogika/mapper package.
+ *
+ * (c) Priyadi Iman Nurcahyo <https://rekalogika.dev>
+ *
+ * For the full copyright and license information, please view the LICENSE file
+ * that was distributed with this source code.
+ */
+
+namespace Rekalogika\Mapper\Tests\Fixtures\Basic;
+
+class PersonWithDog
+{
+    public bool $dogSetterCalled = false;
+
+    public function __construct(
+        private string $name,
+        private ?Dog $dog = null,
+    ) {}
+
+    public function getName(): string
+    {
+        return $this->name;
+    }
+
+    public function setName(string $name): void
+    {
+        $this->name = $name;
+    }
+
+    public function getDog(): ?Dog
+    {
+        return $this->dog;
+    }
+
+    public function setDog(?Dog $dog): void
+    {
+        $this->dogSetterCalled = true;
+        $this->dog = $dog;
+    }
+}